Check authentication status before loading any data
Solved error where the app will try to fetch data regardless of auth status, now frontend checks to see if user is authenticated from the server first by checking 'user.is_authenitcated' and returning according response to frontend, which will make the frontend set logged in status from this response. Before it used to check auth status just by 'accessToken', but this method is incorrect because this does not consider the authentication from the server side, which is where all the data will be fetched from.